pyproject.toml
[tool.poetry]
name = "api-copiloto"
version = "0.1.0"
description = ""
authors = ["Guilherme Brida <guilherme.brida@crearesistemas.com.br>"]
readme = "README.md"
packages = [{include = "api_copiloto"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
python-decouple = "^3.8"
requests = "^2.31.0"
requests-mock = "^1.11.0"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
flask-testing = "^0.8.1"
pytest = "^7.4.0"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
